+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 - Fixed the implementation of `Display` for all types: now the formatting arguments are properly used and the result is identical to the one produced by [`f32`] or [`f64`].
+- Implemented `LowerExp` and `UpperExp` for all types.
 
 ## [0.2.0]
 ### Added